% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download datasets}
\usage{
download(
  pkgname = "MASS",
  dataset = "Boston",
  source = "https://cran.r-universe.dev/"
)
}
\arguments{
\item{pkgname}{a string; R package name}

\item{dataset}{a string; dataset name}

\item{source}{a string; package location (address)}
}
\value{
A data frame
}
\description{
Download datasets
}
\examples{

df <- mlsauce::download(pkgname="MASS", dataset="Boston", source="https://cran.r-universe.dev/")
print(df)

}
