% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazybooster.R
\name{LazyBoostingClassifier}
\alias{LazyBoostingClassifier}
\title{LazyBoostingClassifier}
\usage{
LazyBoostingClassifier(
  verbose = 0,
  ignore_warnings = TRUE,
  custom_metric = NULL,
  predictions = FALSE,
  sort_by = "Accuracy",
  random_state = 42L,
  estimators = "all",
  preprocess = FALSE,
  n_jobs = NULL,
  venv_path = "./venv",
  ...
)
}
\arguments{
\item{verbose:}{int, progress bar (yes = 1) or not (no = 0) (currently).}

\item{ignore_warnings:}{bool, ignore warnings.}

\item{custom_metric:}{function, custom metric.}

\item{predictions:}{bool, return predictions.}

\item{sort_by:}{str, sort by metric.}

\item{random_state:}{int, random state.}

\item{estimators:}{str, estimators to use. List of names for custom, or just 'all'.}

\item{preprocess:}{bool, preprocess data or not.}

\item{n_jobs:}{int, number of jobs.}
}
\value{
LazyBoostingClassifier object
}
\description{
Lazy Generic Boosting Classifier (AutoML Hold-out set validation)
}
\examples{

# See https://thierrymoudiki.github.io/blog/2024/10/14/r/genericboosting-r for advanced examples

library(mlsauce)
library(datasets)

data(iris)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

n <- dim(X)[1]
p <- dim(X)[2]

set.seed(21341)
train_index <- sample(x = 1:n, size = floor(0.8*n), replace = TRUE)
test_index <- -train_index

X_train <- as.matrix(X[train_index, ])
y_train <- as.integer(y[train_index])
X_test <- as.matrix(X[test_index, ])
y_test <- as.integer(y[test_index])

obj <- LazyBoostingClassifier(verbose=0, ignore_warnings=TRUE,
                              custom_metric=NULL, preprocess=FALSE)

obj$fit(X_train, X_test, y_train, y_test)

}
