% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ms}
\alias{ms}
\title{Python module mlsauce}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
ms
}
\description{
This is the Python \code{mlsauce} module imported using reticulate.
}
\keyword{datasets}
