% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{numpy}
\alias{numpy}
\title{Python module numpy}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
numpy
}
\description{
This is the Python \code{numpy} module imported using reticulate.
}
\keyword{datasets}
