% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pandas}
\alias{pandas}
\title{Python module pandas}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
pandas
}
\description{
This is the Python \code{pandas} module imported using reticulate.
}
\keyword{datasets}
