% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sklearn}
\alias{sklearn}
\title{Python module sklearn}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
sklearn
}
\description{
This is the Python \code{sklearn} module imported using reticulate.
}
\keyword{datasets}
